separete duties
